only one process per node creates a machinefile

+++ startup-scripts/gaspi_slurm.sh
@@ -5,17 +5,18 @@ function mysort { for i in ${node_array[@]}; do echo "$i"; done | sort -n; }
 program=$*
 
 export GASPI_SOCKET="$SLURM_LOCALID"
-export GASPI_MFILE="$HOME/machinefile_$GASPI_SOCKET"
+export GASPI_MFILE="$HOME/machinefile_$SLURM_PROCID"
 
 NODES=`srun hostname`
 
 node_array=($NODES)
 
 sorted_array=( $(mysort) )
-
-echo "${sorted_array[@]}" | tr ' ' '\n' > $GASPI_MFILE
-
 export GASPI_MASTER=${sorted_array[0]}
+
+if [ 0 -eq "$GASPI_SOCKET" ];then
+  echo "${sorted_array[@]}" | tr ' ' '\n' > $GASPI_MFILE
+fi
 
 if [ 0 -eq "$SLURM_PROCID" ];then
   export GASPI_TYPE="GASPI_MASTER"
@@ -23,8 +24,11 @@ if [ 0 -eq "$SLURM_PROCID" ];then
   echo "<<start application>>"
 else
   export GASPI_TYPE="GASPI_WORKER"
+  echo "<<start worker $SLURM_PROCID>>"
 fi
 
 $program
 
-rm $GASPI_MFILE
+if [ 0 -eq "$GASPI_SOCKET" ];then
+  rm $GASPI_MFILE
+fi
